Update Toy Musical link and remove Toy Musical 3
I updated the Toy Musical link so that you can now download the Toy Musical 1 and 2 PMS packs.

I have to remove Toy Musical 3, as the files are actually not PMS, they are in a propietary file format called .n2s in which they need .n2x to work. They were only made to be played on its bundled player which means the music cannot be played on neither raja or LR2.